requirements.txt
hvac==0.11.0
PyYAML==5.4.1
cryptography==3.4.8
